[wip] Started porting heatmap to object model
