+++ scanpy/plotting/_anndata.py
@@ -28,7 +28,6 @@ from . import _utils
 from ._utils import scatter_base, scatter_group, setup_axes
 from ._utils import ColorLike, _FontWeight, _FontSize, _AxesSubplot
 from ._docs import doc_scatter_basic, doc_show_save_ax, doc_common_plot_args
-from ._groupby_plots import MatrixPlot, DotPlot, StackedViolin
 
 VALID_LEGENDLOCS = {
     'none',
@@ -853,731 +852,6 @@ def clustermap(
         pl.show()
     else:
         return g
-
-
-@_doc_params(show_save_ax=doc_show_save_ax, common_plot_args=doc_common_plot_args)
-def stacked_violin(
-    adata: AnnData,
-    var_names: Union[_VarNames, Mapping[str, _VarNames]],
-    groupby: Optional[str] = None,
-    log: bool = False,
-    use_raw: Optional[bool] = None,
-    num_categories: int = 7,
-    figsize: Optional[Tuple[float, float]] = None,
-    dendrogram: Union[bool, str] = False,
-    gene_symbols: Optional[str] = None,
-    var_group_positions: Optional[Sequence[Tuple[int, int]]] = None,
-    var_group_labels: Optional[Sequence[str]] = None,
-    standard_scale: Optional[Literal['var', 'obs']] = None,
-    var_group_rotation: Optional[float] = None,
-    layer: Optional[str] = None,
-    stripplot: bool = False,
-    jitter: Union[float, bool] = False,
-    size: int = 1,
-    scale: Literal['area', 'count', 'width'] = 'width',
-    order: Optional[Sequence[str]] = None,
-    swap_axes: bool = False,
-    show: Optional[bool] = None,
-    save: Union[bool, str, None] = None,
-    row_palette: str = 'muted',
-    ax: Optional[_AxesSubplot] = None,
-    **kwds,
-):
-    """\
-    Stacked violin plots.
-
-    Makes a compact image composed of individual violin plots
-    (from :func:`~seaborn.violinplot`) stacked on top of each other.
-    Useful to visualize gene expression per cluster.
-
-    Wraps :func:`seaborn.violinplot` for :class:`~anndata.AnnData`.
-
-    Parameters
-    ----------
-    {common_plot_args}
-    stripplot
-        Add a stripplot on top of the violin plot.
-        See :func:`~seaborn.stripplot`.
-    jitter
-        Add jitter to the stripplot (only when stripplot is True)
-        See :func:`~seaborn.stripplot`.
-    size
-        Size of the jitter points.
-    order
-        Order in which to show the categories. Note: if `dendrogram=True`
-        the categories order will be given by the dendrogram and `order`
-        will be ignored.
-    scale
-        The method used to scale the width of each violin.
-        If 'width' (the default), each violin will have the same width.
-        If 'area', each violin will have the same area.
-        If 'count', a violin’s width corresponds to the number of observations.
-    row_palette
-        The row palette determines the colors to use for the stacked violins.
-        The value should be a valid seaborn or matplotlib palette name
-        (see :func:`~seaborn.color_palette`).
-        Alternatively, a single color name or hex value can be passed,
-        e.g. `'red'` or `'#cc33ff'`.
-    standard_scale
-        Whether or not to standardize a dimension between 0 and 1,
-        meaning for each variable or observation,
-        subtract the minimum and divide each by its maximum.
-    swap_axes
-         By default, the x axis contains `var_names` (e.g. genes) and the y axis the `groupby` categories.
-         By setting `swap_axes` then x are the `groupby` categories and y the `var_names`. When swapping
-         axes var_group_positions are no longer used
-    {show_save_ax}
-    **kwds
-        Are passed to :func:`~seaborn.violinplot`.
-
-    Returns
-    -------
-    List of :class:`~matplotlib.axes.Axes`
-
-    Examples
-    -------
-    >>> import scanpy as sc
-    >>> adata = sc.datasets.pbmc68k_reduced()
-    >>> markers = ['C1QA', 'PSAP', 'CD79A', 'CD79B', 'CST3', 'LYZ']
-    >>> sc.pl.stacked_violin(adata, markers, groupby='bulk_labels', dendrogram=True)
-
-    Using var_names as dict:
-    >>> markers = {{'T-cell': 'CD3D', 'B-cell': 'CD79A', 'myeloid': 'CST3'}}
-    >>> sc.pl.stacked_violin(adata, markers, groupby='bulk_labels', dendrogram=True)
-
-    Get StackeViolin object for fine tuning
-    >>> vp = sc.pl.dotplot(adata, markers, 'bulk_labels', show=False)
-    >>> vp.add_totals().style(ylim=(0,5)).show()
-
-    The axes used can be obtained using the get_axes() method
-    >>> axes_dict = vp.get_axes()
-
-    See also
-    --------
-    rank_genes_groups_stacked_violin: to plot marker genes identified using the :func:`~scanpy.tl.rank_genes_groups` function.
-    """
-
-    dp = StackedViolin(
-            adata,
-            var_names,
-            groupby=groupby,
-            use_raw=use_raw,
-            log=log,
-            num_categories=num_categories,
-            standard_scale=standard_scale,
-            figsize=figsize,
-            gene_symbols=gene_symbols,
-            var_group_positions=var_group_positions,
-            var_group_labels=var_group_labels,
-            var_group_rotation=var_group_rotation,
-            layer=layer,
-            ** kwds,
-    )
-
-    if dendrogram:
-        dp.add_dendrogram(dendrogram_key=dendrogram)
-    if swap_axes:
-        dp.swap_axes()
-    dp = dp.style(stripplot=stripplot, jitter=jitter, jitter_size=size,
-                  row_palette=row_palette, scale=scale)
-    if show is not False or save:
-        return dp.show(show=show, save=save)
-    else:
-        return dp
-
-
-@_doc_params(show_save_ax=doc_show_save_ax, common_plot_args=doc_common_plot_args)
-def heatmap(
-    adata: AnnData,
-    var_names: Union[_VarNames, Mapping[str, _VarNames]],
-    groupby: Optional[str] = None,
-    use_raw: Optional[bool] = None,
-    log: bool = False,
-    num_categories: int = 7,
-    dendrogram: Union[bool, str] = False,
-    gene_symbols: Optional[str] = None,
-    var_group_positions: Optional[Sequence[Tuple[int, int]]] = None,
-    var_group_labels: Optional[Sequence[str]] = None,
-    var_group_rotation: Optional[float] = None,
-    layer: Optional[str] = None,
-    standard_scale: Optional[Literal['var', 'obs']] = None,
-    swap_axes: bool = False,
-    show_gene_labels: Optional[bool] = None,
-    show: Optional[bool] = None,
-    save: Union[str, bool, None] = None,
-    figsize: Optional[Tuple[float, float]] = None,
-    **kwds,
-):
-    """\
-    Heatmap of the expression values of genes.
-
-    If `groupby` is given, the heatmap is ordered by the respective group. For
-    example, a list of marker genes can be plotted, ordered by clustering. If
-    the `groupby` observation annotation is not categorical the observation
-    annotation is turned into a categorical by binning the data into the number
-    specified in `num_categories`.
-
-    Parameters
-    ----------
-    {common_plot_args}
-    standard_scale
-        Whether or not to standardize that dimension between 0 and 1, meaning for each variable or observation,
-        subtract the minimum and divide each by its maximum.
-    swap_axes
-         By default, the x axis contains `var_names` (e.g. genes) and the y axis the `groupby`
-         categories (if any). By setting `swap_axes` then x are the `groupby` categories and y the `var_names`.
-    show_gene_labels
-         By default gene labels are shown when there are 50 or less genes. Otherwise the labels are removed.
-    {show_save_ax}
-    **kwds
-        Are passed to :func:`matplotlib.pyplot.imshow`.
-
-    Returns
-    -------
-    List of :class:`~matplotlib.axes.Axes`
-
-    Examples
-    -------
-    >>> import scanpy as sc
-    >>> adata = sc.datasets.pbmc68k_reduced()
-    >>> markers = ['C1QA', 'PSAP', 'CD79A', 'CD79B', 'CST3', 'LYZ']
-    >>> sc.pl.heatmap(adata, markers, groupby='bulk_labels', dendrogram=True, swap_axes=True)
-
-    Using var_names as dict:
-
-    >>> markers = {{'T-cell': 'CD3D', 'B-cell': 'CD79A', 'myeloid': 'CST3'}}
-    >>> sc.pl.heatmap(adata, markers, groupby='bulk_labels', dendrogram=True)
-
-    See also
-    --------
-    rank_genes_groups_heatmap: to plot marker genes identified using the :func:`~scanpy.tl.rank_genes_groups` function.
-    """
-    if use_raw is None and adata.raw is not None:
-        use_raw = True
-
-    var_names, var_group_labels, var_group_positions = _check_var_names_type(
-        var_names, var_group_labels, var_group_positions
-    )
-
-    categories, obs_tidy = _prepare_dataframe(
-        adata,
-        var_names,
-        groupby,
-        use_raw,
-        log,
-        num_categories,
-        gene_symbols=gene_symbols,
-        layer=layer,
-    )
-
-    if standard_scale == 'obs':
-        obs_tidy = obs_tidy.sub(obs_tidy.min(1), axis=0)
-        obs_tidy = obs_tidy.div(obs_tidy.max(1), axis=0).fillna(0)
-    elif standard_scale == 'var':
-        obs_tidy -= obs_tidy.min(0)
-        obs_tidy = (obs_tidy / obs_tidy.max(0)).fillna(0)
-    elif standard_scale is None:
-        pass
-    else:
-        logg.warning('Unknown type for standard_scale, ignored')
-
-    if groupby is None or len(categories) <= 1:
-        categorical = False
-        # dendrogram can only be computed  between groupby categories
-        dendrogram = False
-    else:
-        categorical = True
-        # get categories colors:
-        if groupby + "_colors" in adata.uns:
-            groupby_colors = adata.uns[groupby + "_colors"]
-        else:
-            groupby_colors = None
-
-    if dendrogram:
-        dendro_data = _reorder_categories_after_dendrogram(
-            adata,
-            groupby,
-            dendrogram,
-            var_names=var_names,
-            var_group_labels=var_group_labels,
-            var_group_positions=var_group_positions,
-        )
-
-        var_group_labels = dendro_data['var_group_labels']
-        var_group_positions = dendro_data['var_group_positions']
-
-        # reorder obs_tidy
-        if dendro_data['var_names_idx_ordered'] is not None:
-            obs_tidy = obs_tidy.iloc[:, dendro_data['var_names_idx_ordered']]
-            var_names = [var_names[x] for x in dendro_data['var_names_idx_ordered']]
-
-        obs_tidy.index = obs_tidy.index.reorder_categories(
-            [categories[x] for x in dendro_data['categories_idx_ordered']],
-            ordered=True,
-        )
-
-        # reorder groupby colors
-        if groupby_colors is not None:
-            groupby_colors = [
-                groupby_colors[x] for x in dendro_data['categories_idx_ordered']
-            ]
-
-    if show_gene_labels is None:
-        if len(var_names) <= 50:
-            show_gene_labels = True
-        else:
-            show_gene_labels = False
-            logg.warning(
-                'Gene labels are not shown when more than 50 genes are visualized. '
-                'To show gene labels set `show_gene_labels=True`'
-            )
-    if categorical:
-        obs_tidy = obs_tidy.sort_index()
-
-    colorbar_width = 0.2
-
-    if not swap_axes:
-        # define a layout of 2 rows x 4 columns
-        # first row is for 'brackets' (if no brackets needed, the height of this row is zero)
-        # second row is for main content. This second row is divided into three axes:
-        #   first ax is for the categories defined by `groupby`
-        #   second ax is for the heatmap
-        #   third ax is for the dendrogram
-        #   fourth ax is for colorbar
-
-        dendro_width = 1 if dendrogram else 0
-        groupby_width = 0.2 if categorical else 0
-        if figsize is None:
-            height = 6
-            if show_gene_labels:
-                heatmap_width = len(var_names) * 0.3
-            else:
-                heatmap_width = 8
-            width = heatmap_width + dendro_width + groupby_width
-        else:
-            width, height = figsize
-            heatmap_width = width - (dendro_width + groupby_width)
-
-        if var_group_positions is not None and len(var_group_positions) > 0:
-            # add some space in case 'brackets' want to be plotted on top of the image
-            height_ratios = [0.15, height]
-        else:
-            height_ratios = [0, height]
-
-        width_ratios = [
-            groupby_width,
-            heatmap_width,
-            dendro_width,
-            colorbar_width,
-        ]
-        fig = pl.figure(figsize=(width, height))
-
-        axs = gridspec.GridSpec(
-            nrows=2,
-            ncols=4,
-            width_ratios=width_ratios,
-            wspace=0.15 / width,
-            hspace=0.13 / height,
-            height_ratios=height_ratios,
-        )
-
-        heatmap_ax = fig.add_subplot(axs[1, 1])
-        im = heatmap_ax.imshow(
-            obs_tidy.values, aspect='auto', interpolation="nearest", **kwds
-        )
-        heatmap_ax.set_ylim(obs_tidy.shape[0] - 0.5, -0.5)
-        heatmap_ax.set_xlim(-0.5, obs_tidy.shape[1] - 0.5)
-        heatmap_ax.tick_params(axis='y', left=False, labelleft=False)
-        heatmap_ax.set_ylabel('')
-        heatmap_ax.grid(False)
-
-        # sns.heatmap(obs_tidy, yticklabels="auto", ax=heatmap_ax, cbar_ax=heatmap_cbar_ax, **kwds)
-        if show_gene_labels:
-            heatmap_ax.tick_params(axis='x', labelsize='small')
-            heatmap_ax.set_xticks(np.arange(len(var_names)))
-            heatmap_ax.set_xticklabels(var_names, rotation=90)
-        else:
-            heatmap_ax.tick_params(axis='x', labelbottom=False, bottom=False)
-
-        # plot colorbar
-        _plot_colorbar(im, fig, axs[1, 3])
-
-        if categorical:
-            groupby_ax = fig.add_subplot(axs[1, 0])
-            ticks, labels, groupby_cmap, norm = _plot_categories_as_colorblocks(
-                groupby_ax, obs_tidy, colors=groupby_colors, orientation='left'
-            )
-
-            # add lines to main heatmap
-            line_positions = (
-                np.cumsum(obs_tidy.index.value_counts(sort=False))[:-1] - 0.5
-            )
-            heatmap_ax.hlines(
-                line_positions,
-                -0.73,
-                len(var_names) - 0.5,
-                lw=0.6,
-                zorder=10,
-                clip_on=False,
-            )
-
-        if dendrogram:
-            dendro_ax = fig.add_subplot(axs[1, 2], sharey=heatmap_ax)
-            _plot_dendrogram(
-                dendro_ax, adata, groupby, ticks=ticks, dendrogram_key=dendrogram,
-            )
-
-        # plot group legends on top of heatmap_ax (if given)
-        if var_group_positions is not None and len(var_group_positions) > 0:
-            gene_groups_ax = fig.add_subplot(axs[0, 1], sharex=heatmap_ax)
-            _plot_gene_groups_brackets(
-                gene_groups_ax,
-                group_positions=var_group_positions,
-                group_labels=var_group_labels,
-                rotation=var_group_rotation,
-                left_adjustment=-0.3,
-                right_adjustment=0.3,
-            )
-
-    # swap axes case
-    else:
-        # define a layout of 3 rows x 3 columns
-        # The first row is for the dendrogram (if not dendrogram height is zero)
-        # second row is for main content. This col is divided into three axes:
-        #   first ax is for the heatmap
-        #   second ax is for 'brackets' if any (othwerise width is zero)
-        #   third ax is for colorbar
-
-        dendro_height = 0.8 if dendrogram else 0
-        groupby_height = 0.13 if categorical else 0
-        if figsize is None:
-            if show_gene_labels:
-                heatmap_height = len(var_names) * 0.18
-            else:
-                heatmap_height = 4
-            width = 10
-            height = heatmap_height + dendro_height + groupby_height
-        else:
-            width, height = figsize
-            heatmap_height = height - (dendro_height + groupby_height)
-
-        height_ratios = [dendro_height, heatmap_height, groupby_height]
-
-        if var_group_positions is not None and len(var_group_positions) > 0:
-            # add some space in case 'brackets' want to be plotted on top of the image
-            width_ratios = [width, 0.14, colorbar_width]
-        else:
-            width_ratios = [width, 0, colorbar_width]
-
-        fig = pl.figure(figsize=(width, height))
-        axs = gridspec.GridSpec(
-            nrows=3,
-            ncols=3,
-            wspace=0.25 / width,
-            hspace=0.3 / height,
-            width_ratios=width_ratios,
-            height_ratios=height_ratios,
-        )
-
-        # plot heatmap
-        heatmap_ax = fig.add_subplot(axs[1, 0])
-
-        im = heatmap_ax.imshow(
-            obs_tidy.T.values, aspect='auto', interpolation="nearest", **kwds
-        )
-        heatmap_ax.set_xlim(0, obs_tidy.shape[0])
-        heatmap_ax.set_ylim(obs_tidy.shape[1] - 0.5, -0.5)
-        heatmap_ax.tick_params(axis='x', bottom=False, labelbottom=False)
-        heatmap_ax.set_xlabel('')
-        heatmap_ax.grid(False)
-        if show_gene_labels:
-            heatmap_ax.tick_params(axis='y', labelsize='small', length=1)
-            heatmap_ax.set_yticks(np.arange(len(var_names)))
-            heatmap_ax.set_yticklabels(var_names, rotation=0)
-        else:
-            heatmap_ax.tick_params(axis='y', labelleft=False, left=False)
-
-        if categorical:
-            groupby_ax = fig.add_subplot(axs[2, 0])
-            ticks, labels, groupby_cmap, norm = _plot_categories_as_colorblocks(
-                groupby_ax, obs_tidy, colors=groupby_colors, orientation='bottom',
-            )
-            # add lines to main heatmap
-            line_positions = (
-                np.cumsum(obs_tidy.index.value_counts(sort=False))[:-1] - 0.5
-            )
-            heatmap_ax.vlines(
-                line_positions,
-                -0.5,
-                len(var_names) + 0.35,
-                lw=0.6,
-                zorder=10,
-                clip_on=False,
-            )
-
-        if dendrogram:
-            dendro_ax = fig.add_subplot(axs[0, 0], sharex=heatmap_ax)
-            _plot_dendrogram(
-                dendro_ax,
-                adata,
-                groupby,
-                dendrogram_key=dendrogram,
-                ticks=ticks,
-                orientation='top',
-            )
-
-        # plot group legends next to the heatmap_ax (if given)
-        if var_group_positions is not None and len(var_group_positions) > 0:
-            gene_groups_ax = fig.add_subplot(axs[1, 1])
-            arr = []
-            for idx, pos in enumerate(var_group_positions):
-                arr += [idx] * (pos[1] + 1 - pos[0])
-
-            gene_groups_ax.imshow(
-                np.matrix(arr).T, aspect='auto', cmap=groupby_cmap, norm=norm
-            )
-            gene_groups_ax.axis('off')
-
-        # plot colorbar
-        _plot_colorbar(im, fig, axs[1, 2])
-
-    _utils.savefig_or_show('heatmap', show=show, save=save)
-
-    return axs
-
-
-@_doc_params(show_save_ax=doc_show_save_ax, common_plot_args=doc_common_plot_args)
-def dotplot(
-    adata: AnnData,
-    var_names: Union[_VarNames, Mapping[str, _VarNames]],
-    groupby: Optional[str] = None,
-    use_raw: Optional[bool] = None,
-    log: bool = False,
-    num_categories: int = 7,
-    expression_cutoff: float = 0.0,
-    mean_only_expressed: bool = False,
-    color_map: str = 'Reds',
-    dot_max: Optional[float] = None,
-    dot_min: Optional[float] = None,
-    standard_scale: Literal['var', 'group'] = None,
-    smallest_dot: float = 0.0,
-    color_title: Optional[str] = 'Expression\nlevel in group',
-    size_title: Optional[str] = 'Fraction of cells\nin group (%)',
-    figsize: Optional[Tuple[float, float]] = None,
-    dendrogram: Union[bool, str] = False,
-    gene_symbols: Optional[str] = None,
-    var_group_positions: Optional[Sequence[Tuple[int, int]]] = None,
-    var_group_labels: Optional[Sequence[str]] = None,
-    var_group_rotation: Optional[float] = None,
-    layer: Optional[str] = None,
-    swap_axes: Optional[bool] = False,
-    dot_color_df: Optional[pd.DataFrame] = None,
-    show: Optional[bool] = None,
-    save: Union[str, bool, None] = None,
-    **kwds,
-) -> DotPlot:
-    """\
-    Makes a *dot plot* of the expression values of `var_names`.
-
-    For each var_name and each `groupby` category a dot is plotted.
-    Each dot represents two values: mean expression within each category
-    (visualized by color) and fraction of cells expressing the `var_name` in the
-    category (visualized by the size of the dot). If `groupby` is not given,
-    the dotplot assumes that all data belongs to a single category.
-
-    .. note::
-       A gene is considered expressed if the expression value in the `adata` (or
-       `adata.raw`) is above the specified threshold which is zero by default.
-
-    An example of dotplot usage is to visualize, for multiple marker genes,
-    the mean value and the percentage of cells expressing the gene
-    accross multiple clusters.
-
-    Parameters
-    ----------
-    {common_plot_args}
-    expression_cutoff
-        Expression cutoff that is used for binarizing the gene expression and
-        determining the fraction of cells expressing given genes. A gene is
-        expressed only if the expression value is greater than this threshold.
-    mean_only_expressed
-        If True, gene expression is averaged only over the cells
-        expressing the given genes.
-    color_map
-        String denoting matplotlib color map.
-    dot_max
-        If none, the maximum dot size is set to the maximum fraction value found
-        (e.g. 0.6). If given, the value should be a number between 0 and 1.
-        All fractions larger than dot_max are clipped to this value.
-    dot_min
-        If none, the minimum dot size is set to 0. If given,
-        the value should be a number between 0 and 1.
-        All fractions smaller than dot_min are clipped to this value.
-    standard_scale
-        Whether or not to standardize that dimension between 0 and 1,
-        meaning for each variable or group,
-        subtract the minimum and divide each by its maximum.
-    smallest_dot
-        If none, the smallest dot has size 0.
-        All expression levels with `dot_min` are plotted with this size.
-
-    {show_save_ax}
-    **kwds
-        Are passed to :func:`matplotlib.pyplot.scatter`.
-
-    Returns
-    -------
-    if `show` is False, returns a DotPlot object
-
-    Examples
-    -------
-    >>> import scanpy as sc
-    >>> adata = sc.datasets.pbmc68k_reduced()
-    >>> markers = ['C1QA', 'PSAP', 'CD79A', 'CD79B', 'CST3', 'LYZ']
-    >>> sc.pl.dotplot(adata, markers, groupby='bulk_labels', dendrogram=True)
-
-    Using var_names as dict:
-    >>> markers = {{'T-cell': 'CD3D', 'B-cell': 'CD79A', 'myeloid': 'CST3'}}
-    >>> sc.pl.dotplot(adata, markers, groupby='bulk_labels', dendrogram=True)
-
-    Get DotPlot object for fine tuning
-    >>> dp = sc.pl.dotplot(adata, markers, 'bulk_labels', show=False)
-    >>> dp.add_totals().style(dot_edge_color='black', dot_edge_lw=0.5).show()
-
-    The axes used can be obtained using the get_axes() method
-    >>> axes_dict = dp.get_axes()
-
-    See also
-    --------
-    :func:`~scanpy.pl.rank_genes_groups_dotplot`: to plot marker genes identified using the
-    :func:`~scanpy.tl.rank_genes_groups` function.
-    """
-
-    dp = DotPlot(adata,
-            var_names,
-            groupby,
-            use_raw=use_raw,
-            log=log,
-            num_categories=num_categories,
-            expression_cutoff=expression_cutoff,
-            mean_only_expressed=mean_only_expressed,
-            standard_scale=standard_scale,
-            figsize=figsize,
-            gene_symbols=gene_symbols,
-            var_group_positions=var_group_positions,
-            var_group_labels=var_group_labels,
-            var_group_rotation=var_group_rotation,
-            layer=layer,
-            dot_color_df=dot_color_df,
-            ** kwds,
-    )
-
-    if dendrogram:
-        dp.add_dendrogram(dendrogram_key=dendrogram)
-    if swap_axes:
-        dp.swap_axes()
-
-    dp = dp.style(cmap=color_map, dot_max=dot_max, dot_min=dot_min, smallest_dot=smallest_dot)
-
-    if show is not False or save:
-        return dp.show(show=show, save=save)
-    else:
-        return dp
-
-
-@_doc_params(show_save_ax=doc_show_save_ax, common_plot_args=doc_common_plot_args)
-def matrixplot(
-    adata: AnnData,
-    var_names: Union[_VarNames, Mapping[str, _VarNames]],
-    groupby: Optional[str] = None,
-    use_raw: Optional[bool] = None,
-    log: bool = False,
-    num_categories: int = 7,
-    figsize: Optional[Tuple[float, float]] = None,
-    dendrogram: Union[bool, str] = False,
-    plot_totals: Union[bool, str] = False,
-    gene_symbols: Optional[str] = None,
-    var_group_positions: Optional[Sequence[Tuple[int, int]]] = None,
-    var_group_labels: Optional[Sequence[str]] = None,
-    var_group_rotation: Optional[float] = None,
-    layer: Optional[str] = None,
-    standard_scale: Literal['var', 'group'] = None,
-    values_df: Optional[pd.DataFrame] = None,
-    swap_axes: bool = False,
-    show: Optional[bool] = None,
-    save: Union[str, bool, None] = None,
-    **kwds,
-) -> MatrixPlot:
-    """\
-    Creates a heatmap of the mean expression values per cluster of each var_names
-    If groupby is not given, the matrixplot assumes that all data belongs to a single
-    category.
-
-    Parameters
-    ----------
-    {common_plot_args}
-    standard_scale
-        Whether or not to standardize that dimension between 0 and 1, meaning for each variable or group,
-        subtract the minimum and divide each by its maximum.
-    {show_save_ax}
-    **kwds
-        Are passed to :func:`matplotlib.pyplot.pcolor`.
-
-    Returns
-    -------
-    if `show` is False, returns a MatrixPlot object
-
-    Examples
-    --------
-    >>> import scanpy as sc
-    >>> adata = sc.datasets.pbmc68k_reduced()
-    >>> markers = ['C1QA', 'PSAP', 'CD79A', 'CD79B', 'CST3', 'LYZ']
-    >>> sc.pl.matrixplot(adata, markers, groupby='bulk_labels', dendrogram=True)
-
-    Using var_names as dict:
-    >>> markers = {{'T-cell': 'CD3D', 'B-cell': 'CD79A', 'myeloid': 'CST3'}}
-    >>> sc.pl.matrixplot(adata, markers, groupby='bulk_labels', dendrogram=True)
-
-    Get Matrix object for fine tuning
-    >>> mp = sc.pl.matrix(adata, markers, 'bulk_labels', show=False)
-    >>> mp.add_totals().style(edge_color='black').show()
-
-    The axes used can be obtained using the get_axes() method
-    >>> axes_dict = mp.get_axes()
-
-    See also
-    --------
-    :func:`~scanpy.pl.rank_genes_groups_matrixplot`: to plot marker genes identified using the
-    :func:`~scanpy.tl.rank_genes_groups` function.
-    """
-
-    dp = MatrixPlot(adata,
-            var_names,
-            groupby=groupby,
-            use_raw=use_raw,
-            log=log,
-            num_categories=num_categories,
-            standard_scale=standard_scale,
-            figsize=figsize,
-            gene_symbols=gene_symbols,
-            var_group_positions=var_group_positions,
-            var_group_labels=var_group_labels,
-            var_group_rotation=var_group_rotation,
-            layer=layer,
-            values_df=values_df,
-            ** kwds,
-    )
-
-    if dendrogram:
-        dp.add_dendrogram(dendrogram_key=dendrogram)
-    if swap_axes:
-        dp.swap_axes()
-
-    dp = dp.style(cmap=kwds.get('cmap'))
-    if show is not False or save:
-        return dp.show(show=show, save=save)
-    else:
-        return dp
 
 
 @_doc_params(show_save_ax=doc_show_save_ax, common_plot_args=doc_common_plot_args)
